+++ CHANGELOG.md
@@ -12,12 +12,15 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
-- [config]: Added option "postfix" to postfix the matter serial number to avoid eventual collisionis with other devices on the network (default empty string).
+- [config]: Added option "postfix" to postfix the matter serial number to avoid eventual collisions with other devices on the network (default empty string).
 - [config]: Added option "failsafeCount" to avoid to start the bridge when some network issue prevents to load all devices (default 0 = disabled).
 
 ### Verified
 
-- [shelly]: Verified shellyprodm1pm with firmware v. 1.4.2 and added Jest test (this needs to be calibrated to enable level control).
+- [shelly]: Verified shellyprodm1pm with firmware v. 1.4.2 and added Jest test (this device needs to be calibrated to enable level control).
+- [shelly]: Verified shelly1g3 with firmware v. 1.4.2 and added Jest test.
+- [shelly]: Verified shelly1pmg3 with firmware v. 1.4.2 and added Jest test.
+
 
 ### Changed